Fri, Mar 18, 2022, 11:27 PM +00:00

+++ extra/extra4.c
@@ -31,5 +31,5 @@ float average(int length, int array[])
     {
         sum = sum + array[i];
     }
-    return sum / n;
+    return sum / (int) n;
 }
